+++ config/autoload.php
@@ -2,9 +2,9 @@
 
 /**
  * news_categories extension for Contao Open Source CMS
- * 
+ *
  * Copyright (C) 2013 Codefog
- * 
+ *
  * @package news_categories
  * @link    http://codefog.pl
  * @author  Webcontext <http://webcontext.com>
@@ -24,6 +24,9 @@ ClassLoader::addNamespace('NewsCategories');
  */
 ClassLoader::addClasses(array
 (
+	// Classes
+	'NewsCategories\News'                 => 'system/modules/news_categories/classes/News.php',
+
 	// Models
 	'NewsCategories\NewsCategoryModel'    => 'system/modules/news_categories/models/NewsCategoryModel.php',
 	'NewsCategories\NewsModel'            => 'system/modules/news_categories/models/NewsModel.php',