+++ languages/de/tl_news_feed.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * news_categories extension for Contao Open Source CMS
+ *
+ * Copyright (C) 2013 Codefog
+ *
+ * @package news_categories
+ * @link    http://codefog.pl
+ * @author  Webcontext <http://webcontext.com>
+ * @author  Kamil Kuzminski <kamil.kuzminski@codefog.pl>
+ * @license LGPL
+ */
+
+
+/**
+ * Fields
+ */
+$GLOBALS['TL_LANG']['tl_news_feed']['categories'] = array('Kategorien', 'Eine oder mehrere Kategorien auswählen.');