+++ languages/fr/tl_news_feed.php
@@ -1,0 +1,20 @@
+<?php
+
+/**
+ * news_categories extension for Contao Open Source CMS
+ *
+ * Copyright (C) 2013 Codefog
+ *
+ * @package news_categories
+ * @link    http://codefog.pl
+ * @author  Webcontext <http://webcontext.com>
+ * @author  Kamil Kuzminski <kamil.kuzminski@codefog.pl>
+ * @author  Lionel Maccaud <https://github.com/lionel-m>
+ * @license LGPL
+ */
+
+
+/**
+ * Fields
+ */
+$GLOBALS['TL_LANG']['tl_news_feed']['categories'] = array('Catégories', 'Ici, vous pouvez sélectionner une ou plusieurs catégories d\'actualité.');