+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Version 1.1.0 stable (2013-??-??)
 
 ### Improved
 - Added the Italian translation (thanks to Marco Damian)
+- Added the categories to RSS feeds (see #4)
 
 
 Version 1.0.9 stable (2013-08-10)